requirements.txt
click==6.7
SQLAlchemy==1.3.10
SQLAlchemy-Searchable==0.10.6
SQLAlchemy-Utils==0.34.2
python-dotenv==0.16.0
simplejson==3.17.2
watchdog==2.0.2
passlib==1.7.1
funcy==1.15
redis==3.5.3
PyYAML==5.4.1
psycopg2==2.8.3
PyMySQL==1.0.2
PyJWT==1.7.1
Flask==1.1.1
Flask-Cors==3.0.9
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.0
Flask-Migrate==2.5.2
flask-mail==0.9.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1